Set image tag appropriately
