+++ .github/workflows/kubernetes_test.yaml
@@ -35,6 +35,7 @@ jobs:
       - name: 'QHUB_GH_BRANCH set for PR'
         run: |
           echo "QHUB_GH_BRANCH=${GITHUB_SHA}" >> "$GITHUB_ENV"
+          echo "QHUB_IMAGE_TAG=main" >> "$GITHUB_ENV"
           echo "GITHUB_BASE_REF: ${GITHUB_BASE_REF}"
           echo "GITHUB_HEAD_REF: ${GITHUB_HEAD_REF}"
           echo "GITHUB_REF: ${GITHUB_REF}"
@@ -83,13 +84,6 @@ jobs:
         run: |
           docker info
           docker ps
-      - name: Pull docker images
-        run: |
-          docker pull quay.io/nebari/nebari-jupyterhub:main
-          docker pull quay.io/nebari/nebari-jupyterlab:main
-          docker pull quay.io/nebari/nebari-dask-worker:main
-        env:
-          TAG: ${{ github.sha }}
       - name: List docker images in minikube
         run: |
           docker images
