+++ qhub/provider/cicd/common.py
@@ -1,0 +1,13 @@
+import os
+
+
+def pip_install_qhub(qhub_version: str) -> str:
+    qhub_gh_branch = os.environ.get("QHUB_GH_BRANCH")
+    pip_install = f"pip install qhub=={qhub_version}"
+    # dev branches
+    if len(qhub_version.split(".")) > 3 and qhub_gh_branch:
+        pip_install = (
+            f"pip install git+https://github.com/Quansight/qhub.git@{qhub_gh_branch}"
+        )
+
+    return pip_install
